New feature: find upstream repos, and find backports in them
This adds code to find upstream repos from package repos - for
now it just parses the URL and Source lines and looks for github
URLs, it doesn't support anything else - and then find backport
commits within the upstream repos. For now it identifies a
backport commit the same way Tomas' script does (just looks for
the text "backport of " in the commit message). It also tries
to find the commit it's a backport of (for now, by comparing
commit messages). All of this can be extended and refined. For
now, the results are:

Found 49 backport commits with identifiable source commits!
Found 253 backport commits without identifiable source commits!

Signed-off-by: Adam Williamson <awilliam@redhat.com>